Fixed Discount

+++ src/Accounts.Application/Projects/Dto/ProjectDto.cs
@@ -30,7 +30,7 @@ namespace Accounts.Projects.Dto
 
         public DiscountType? DiscountType { get; set; }
 
-        public double? DiscountValue { get; set; }
+        public decimal? DiscountValue { get; set; }
 
         public double Rate { get; set; }
